# contracts/openapi.yaml
openapi: 3.1.0
...
info:
  title: DevReport AI API
  version: 0.1.0
  description: |
    인증 요청의 IP별 rate limit은 Spring의 `server.forward-headers-strategy=framework`로
    해석된 클라이언트 IP를 사용한다. 배포 시 신뢰된 reverse proxy 또는 ingress만
    `Forwarded`/`X-Forwarded-For`를 정규화해 전달하고, 외부에서 전달된 forwarded header는
    제거해야 한다. Refresh Token은 `refresh_token` HttpOnly 쿠키로만 전달하며,
    Frontend는 인증 요청에 credentials를 포함해야 한다. Backend는 명시된 허용 Origin만
    credentials 요청에 허용하고, 쿠키를 사용하는 인증 요청의 Origin 또는 Referer를 검증한다.
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshTokenCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: HttpOnly·SameSite Refresh Token 쿠키. JavaScript에서 읽을 수 없다.
  schemas:
...
    ReportSummaryResponse:
      type: object
      required: [id, templateId, templateVersion, version, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        templateId:
          type: [string, "null"]
          pattern: '^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$'
        templateVersion:
          type: [integer, "null"]
          minimum: 1
        version:
          type: integer
          minimum: 0
        updatedAt:
          type: string
          format: date-time
    ReportPageResponse:
      type: object
      required: [items, page, size, totalElements, totalPages]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ReportSummaryResponse"
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
          maximum: 100
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
    ReportPresentationSettings:
      type: object
      propertyNames:
        pattern: '^[A-Za-z][A-Za-z0-9_-]{0,63}$'
      additionalProperties:
        type: [string, number, boolean, "null"]
    ReportUpdateRequest:
      type: object
      required: [document, templateId, templateVersion, presentationSettings, expectedVersion]
      properties:
        document:
          $ref: "report-document.schema.json"
        templateId:
          type: [string, "null"]
          pattern: '^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$'
        templateVersion:
          type: [integer, "null"]
          minimum: 1
        presentationSettings:
          $ref: "#/components/schemas/ReportPresentationSettings"
        expectedVersion:
          type: integer
          minimum: 0
      oneOf:
        - properties:
            templateId:
              type: "null"
            templateVersion:
              type: "null"
        - properties:
            templateId:
              type: string
            templateVersion:
              type: integer
...
    ExportIdResponse:
      type: object
      required: [exportId]
      properties:
        exportId:
          type: string
          format: uuid
    ReportExportResponse:
      type: object
      required: [exportId, reportId, status, createdAt]
      properties:
        exportId:
          type: string
          format: uuid
        reportId:
          type: string
          format: uuid
        status:
          type: string
          enum: [PENDING, PROCESSING, COMPLETED, FAILED, EXPIRED]
        size:
          type: [integer, "null"]
          format: int64
        failureCode:
          type: [string, "null"]
        failureMessage:
          type: [string, "null"]
        expiresAt:
          type: [string, "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: [string, "null"]
          format: date-time
        completedAt:
          type: [string, "null"]
          format: date-time
    ReportRenderDataResponse:
      type: object
      required: [exportId, reportId, projectId, reportVersion, document, templateId, templateVersion,
        presentationSettings, imageFileIds]
      properties:
        exportId:
          type: string
          format: uuid
        reportId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        reportVersion:
          type: integer
          format: int64
          minimum: 0
        document:
          $ref: "report-document.schema.json"
        templateId:
          type: string
          pattern: '^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$'
        templateVersion:
          type: integer
          minimum: 1
        presentationSettings:
          $ref: "#/components/schemas/ReportPresentationSettings"
        imageFileIds:
          type: array
          uniqueItems: true
          items:
            type: string
            format: uuid
    ErrorResponse:
      type: object
      required: [code, message, timestamp]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: [object, "null"]
        timestamp:
          type: string
          format: date-time
security:
  - bearerAuth: []
paths:
...
  /api/projects/{projectId}/reports:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: 프로젝트 보고서 목록 조회
      description: 수정 시각과 ID 내림차순으로 보고서 목록을 조회한다.
      parameters:
        - name: page
          in: query
          description: 0부터 시작하는 페이지 번호
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          description: 페이지당 보고서 수
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: 프로젝트 보고서 목록 조회 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportPageResponse"
        "400":
          description: 잘못된 페이지 요청
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 미인증
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 프로젝트가 없거나 소유자가 아님
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/reports/{reportId}:
    get:
      summary: 보고서 조회
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 보고서 조회 성공. 응답의 projectId는 보고서 소속 프로젝트 ID이며 이미지 파일 조회에 사용한다.
          content:
            application/json:
              schema:
                $ref: "report.schema.json"
        "401":
          description: 미인증
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 보고서가 없거나 프로젝트 소유자가 아님 (REPORT_NOT_FOUND)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      summary: 보고서 수정
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportUpdateRequest"
      responses:
        "200":
          description: 보고서 수정 성공. 응답의 projectId는 보고서 소속 프로젝트 ID이며 이미지 파일 조회에 사용한다.
          content:
            application/json:
              schema:
                $ref: "report.schema.json"
        "400":
          description: ReportDocument·템플릿·버전 요청 오류 (REPORT_DOCUMENT_INVALID, REPORT_PRESENTATION_INVALID, REPORT_VERSION_INVALID)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 미인증
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 보고서가 없거나 프로젝트 소유자가 아님 (REPORT_NOT_FOUND)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: expectedVersion이 현재 보고서 버전과 다름 (REPORT_VERSION_CONFLICT)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/reports/{reportId}/exports:
    post:
      summary: 보고서 PDF 생성 요청
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "202":
          description: PDF 생성 작업 접수
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportIdResponse"
        "401":
          description: 미인증
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 보고서가 없거나 프로젝트 소유자가 아님 (REPORT_NOT_FOUND)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: 보고서 템플릿이 선택되지 않음 (REPORT_TEMPLATE_NOT_SELECTED)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: PDF 출력 URL이 설정되지 않았거나 올바르지 않음 (EXPORT_PRINT_URL_NOT_CONFIGURED)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: 사용자별 PDF 일일·동시 quota 초과 (PDF_DAILY_LIMIT_EXCEEDED, PDF_CONCURRENCY_LIMIT_EXCEEDED)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/report-exports/{exportId}:
    get:
      summary: PDF 생성 상태 조회
      parameters:
        - name: exportId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: PDF 생성 상태
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportExportResponse"
        "401":
          description: 미인증
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 작업이 없거나 프로젝트 소유자가 아님 (EXPORT_NOT_FOUND)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/report-exports/{exportId}/render-data:
    get:
      summary: PDF 출력용 보고서 스냅샷 조회
      security: []
      parameters:
        - name: exportId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Render-Token
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: PDF 출력용 요청 시점 스냅샷
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportRenderDataResponse"
        "404":
          description: 유효하지 않거나 만료된 출력 토큰 또는 작업 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/report-exports/{exportId}/files/{fileId}:
    get:
      summary: PDF 출력용 이미지 조회
      security: []
      parameters:
        - name: exportId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: fileId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Render-Token
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 보고서 스냅샷에 포함된 이미지 파일
          headers:
            Content-Disposition:
              schema:
                type: string
          content:
            image/png:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
        "404":
          description: 유효하지 않거나 만료된 출력 토큰·작업·이미지
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/report-exports/{exportId}/download:
    get:
      summary: 생성된 PDF 다운로드
      parameters:
        - name: exportId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: PDF 파일
          headers:
            Content-Disposition:
              schema:
                type: string
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "401":
          description: 미인증
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 작업·파일이 없거나 프로젝트 소유자가 아님 (EXPORT_NOT_FOUND, EXPORT_FILE_NOT_FOUND)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: PDF가 준비되지 않았거나 생성 실패 (EXPORT_NOT_READY, EXPORT_FAILED)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "410":
          description: PDF 다운로드 기간 만료 (EXPORT_EXPIRED)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"